requirements.txt
numpy
opencv-python
opencv-python-headless
pillow
tensorflow